[JBWS-4133] Use reflection to check if a Principal is a Group
This approach relies on the fact that the first Principal in a subject will represent the caller's identity.
The code already assumes that the first non-Group Principal is that, so this is not a new assumption.
Code that populates a Subject that cannot use Group must ensure that whatever principal represents the security identity gets added first.